requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
contourpy==1.3.1
cryptography==43.0.0
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
Django==5.0.7
django-allauth==64.0.0
django-cors-headers==4.4.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
exponent_server_sdk==2.1.0
faiss-cpu==1.9.0
filelock==3.16.1
fonttools==4.55.0
frozenlist==1.4.1
fsspec==2024.10.0
git-filter-repo==2.45.0
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-dialogflow==2.31.0
googleapis-common-protos==1.63.2
greenlet==3.1.1
grpcio==1.66.0
grpcio-status==1.66.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.7
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
JPype1==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
konlpy==0.6.0
langchain==0.3.3
langchain-community==0.3.2
langchain-core==0.3.11
langchain-openai==0.2.2
langchain-text-splitters==0.3.0
langsmith==0.1.135
lxml==5.3.0
MarkupSafe==3.0.2
marshmallow==3.22.0
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==1.26.4
openai==1.51.2
openpyxl==3.1.5
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pillow==10.4.0
propcache==0.2.0
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
PyJWT==2.9.0
PyMuPDF==1.24.11
pyparsing==3.1.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
qrcode==7.4.2
redis==5.0.8
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.1
six==1.16.0
slack_sdk==3.33.4
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.1
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.1
torch==2.5.1
tqdm==4.66.5
transformers==4.46.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.15.4
